+++ R/misc_yhat.R
@@ -73,6 +73,11 @@ yhat.svm <- function(X.model, newdata, ...) {
 yhat.gbm <- function(X.model, newdata, ...) {
   n.trees <- X.model$n.trees
   response <- predict(X.model, newdata = newdata, n.trees = n.trees, type = "response")
+  #gbm returns and 3D array for multilabel classif
+  if(length(dim(response)) > 2){
+    response <- response[,,1]
+  }
+  response
 }
 
 
@@ -90,6 +95,10 @@ yhat.cv.glmnet <- function(X.model, newdata, ...) {
   }
   if (!is.null(X.model$glmnet.fit$classnames)) {
     pred <- predict(X.model, newdata, type = "response", s = X.model$lambda[length(X.model$lambda)])
+    #glmnet returns and 3D array for multilabel classif
+    if(length(dim(pred)) > 2){
+      return(pred[,,1])
+    }
     if (ncol(pred) == 1) {
       return(as.numeric(pred))
     }
@@ -110,6 +119,10 @@ yhat.glmnet <- function(X.model, newdata, ...) {
   }
   if (!is.null(X.model$classnames)) {
     pred <- predict(X.model, newdata, type = "response", s = X.model$lambda[length(X.model$lambda)])
+    #glmnet returns and 3D array for multilabel classif
+    if(length(dim(pred)) > 2){
+      return(pred[,,1])
+    }
   # For binary classifiaction matrix with one column is returned
     if (ncol(pred) == 1) {
       return(as.numeric(pred))
@@ -143,6 +156,7 @@ yhat.ranger <- function(X.model, newdata, ...) {
 yhat.model_fit <- function(X.model, newdata, ...) {
   if (X.model$spec$mode == "classification") {
     response <- as.matrix(predict(X.model, newdata, type = "prob"))
+    colnames(response) <- X.model$lvl
     if (ncol(response) == 2) {
       response <- response[,2]
     }