+++ R/model_performance.R
@@ -10,7 +10,7 @@
 #'
 #' @return An object of the class \code{model_performance}.
 #' @references Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://pbiecek.github.io/ema/}
-#' @importFrom stats median
+#' @importFrom stats median weighted.mean
 #' @export
 #' @examples
 #'  \dontrun{
@@ -53,7 +53,7 @@ model_performance <- function(explainer, ..., cutoff = 0.5) {
   observed <- explainer$y
   # Check since explain could have been run with precalculate = FALSE
   if (is.null(explainer$residuals)) {
-    diff <- explainer$residual_function(explainer$model, explainer$data, observed)
+    diff <- explainer$residual_function(explainer$model, explainer$data, observed, explainer$predict_function)
   } else {
     # changed according to #130
     diff <- explainer$residuals
@@ -70,7 +70,7 @@ model_performance <- function(explainer, ..., cutoff = 0.5) {
       r2 = model_performance_r2(predicted, observed),
       mad = model_performance_mad(predicted, observed)
     )
-  } else {
+  } else if (type == "classification") {
     tp = sum((observed == 1) * (predicted >= cutoff))
     fp = sum((observed == 0) * (predicted >= cutoff))
     tn = sum((observed == 0) * (predicted < cutoff))
@@ -82,6 +82,14 @@ model_performance <- function(explainer, ..., cutoff = 0.5) {
       f1        = model_performance_f1(tp, fp, tn, fn),
       accuracy  = model_performance_accuracy(tp, fp, tn, fn),
       auc       = model_performance_auc(predicted, observed)
+    )
+  } else {
+    measures <- list(
+      micro_F1 = model_performance_micro_f1(predicted, observed),
+      macro_F1 = model_performance_macro_f1(predicted, observed),
+      w_macro_F1 = model_performance_weighted_macro_f1(predicted, observed),
+      accuracy = model_performance_accuracy_multi(predicted, observed),
+      w_macro_auc = model_performance_weighted_macro_auc(predicted, observed)
     )
   }
 
@@ -139,3 +147,60 @@ model_performance_accuracy <- function(tp, fp, tn, fn) {
   (tp + tn)/(tp + fp + tn + fn)
 }
 
+model_performance_macro_f1 <- function(predicted, observed) {
+  predicted_vectorized <- turn_probs_into_vector(predicted)
+  confusion_matrixes <- calculate_confusion_matrixes(predicted_vectorized, observed)
+  f1_scores <- sapply(confusion_matrixes, function(x){
+    model_performance_f1(x$tp, x$fp, x$tn, x$fn)
+  })
+  mean(f1_scores)
+}
+
+model_performance_micro_f1 <- function(predicted, observed) {
+  # For case where each point can be assigned only to one class micro_f1 equals acc
+  model_performance_accuracy_multi(predicted, observed)
+}
+
+model_performance_weighted_macro_f1 <- function(predicted, observed) {
+  predicted_vectorized <- turn_probs_into_vector(predicted)
+  confusion_matrixes <- calculate_confusion_matrixes(predicted_vectorized, observed)
+  f1_scores <- sapply(confusion_matrixes, function(x){
+    model_performance_f1(x$tp, x$fp, x$tn, x$fn)
+  })
+  weighted.mean(f1_scores, prop.table(table(observed))[names(confusion_matrixes)])
+}
+
+model_performance_accuracy_multi <- function(predicted, observed) {
+  predicted_vectorized <- turn_probs_into_vector(predicted)
+  mean(predicted_vectorized == observed)
+}
+
+model_performance_weighted_macro_auc <- function(predicted, observed) {
+  observed <- as.character(observed)
+  auc_scores <- sapply(unique(observed), function(x){
+    model_performance_auc(predicted[,x], as.numeric(observed == x))
+  })
+  weighted.mean(auc_scores, prop.table(table(observed))[unique(observed)])
+}
+
+turn_probs_into_vector <- function(observed) {
+  apply(observed, 1, function(x){
+    colnames(observed)[which.max(x)]
+  })
+}
+
+calculate_confusion_matrixes <- function(predicted, observed) {
+  observed <- as.character(observed)
+  ret <- lapply(unique(observed), function(x){
+    tp <- mean(predicted[predicted == x] == observed[predicted == x])
+    fp <- mean(predicted[predicted == x] != observed[predicted == x])
+    tn <- mean(predicted[predicted != x] == observed[predicted != x])
+    fn <- mean(predicted[predicted != x] != observed[predicted != x])
+    list(tp = ifelse(is.nan(tp), 0, tp),
+         fp = ifelse(is.nan(fp), 0, fp),
+         tn = ifelse(is.nan(tn), 0, tn),
+         fn = ifelse(is.nan(fn), 0, fn))
+  })
+  names(ret) <- unique(observed)
+  ret
+}