+++ NEWS.md
@@ -1,3 +1,10 @@
+DALEX 1.2.2
+----------------------------------------------------------------
+* `DALEX` now fully supports multiclass classification. 
+* `explain()` will use new residual function (1 - true class probability) if multiclass classification is detected.
+* `model_performance()` now support measures for multiclass classification.
+* Remove `ggpubr` from suggests.
+
 DALEX 1.2.1
 ----------------------------------------------------------------
 * fixed tests and WARNINGs on CRAN